+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Release Notes
 
+### 1.0.7
+Replace settings  
+`"bracketPairColorizer.activeScopeBorderStyle"`  
+`"bracketPairColorizer.activeScopeBackgroundColor"`
+
+with `"bracketPairColorizer.activeScopeCSS"`
+
 ### 1.0.6
 Fix `"bracketPairColorizer.highlightActiveScope"` setting not being respected